created update controller, api, component, form

+++ backend/controllers/recipes.js
@@ -1,4 +1,5 @@
 import Recipe from "../models/recipe.js";
+import mongoose from "mongoose";
 
 export const getAllRecipes = async (req, res) => {
   try {
@@ -24,9 +25,11 @@ export const createRecipe = async (req, res) => {
 
 export const getUniqueRecipe = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id))
+    return res.status(404).send(`No post with id: ${id}`);
 
   try {
-    const uniqueRecipe = await RTCIceCandidate.findById(id);
+    const uniqueRecipe = await Recipe.findById(id);
 
     res.status(200).json(uniqueRecipe);
   } catch (error) {
@@ -34,8 +37,21 @@ export const getUniqueRecipe = async (req, res) => {
   }
 };
 
+export const updateRecipe = async (req, res) => {
+  const { id } = req.params;
+  const body = req.body;
+  if (!mongoose.Types.ObjectId.isValid(id))
+    return res.status(404).send(`No post with id: ${id}`);
+
+  const updatedRecipe = await Recipe.findByIdAndUpdate(id, body, { new: true });
+
+  res.status(200).json(updatedRecipe);
+};
+
 export const deleteRecipe = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id))
+    return res.status(404).send(`No post with id: ${id}`);
 
   await Recipe.findByIdAndRemove(id);
